fix share text/data to Forge app - add additional launcher activity to prevent GLES Context loss (when sharing data via intent) - fix passing data from another app to Forge via share text / share file / share deck
