--- forge-gui-android/src/forge/app/Exiter.java
@@ -1,10 +1,0 @@
-package forge.app;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-public class Exiter extends Activity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-    }
-}
